ci(circle): Force the use of the NPM registry
After experiencing some hard to debug authentication issues during the
automatic release of the package, an issue was found were the `yarn`
command uses a different registry than the `npm` command.

By forcing the storage of the `NPM_TOKEN` in the .npmrc and also using
the `npm` command for the semanic release, it is hopefully possible to
automatically release again.