fix: refactor implementation and keep stable interface
This is a refactor proposal that will make it easier to understand and
develop further this library. It uses a configuration for the endpoints
and the `axios` library for the HTTP requests instead of the custom
implementation that relied on the `http` and `https` node modules.

This version keeps the public interface the same to not break backwards
compatibility and has feature parity with the existing implementation.

It also fixes an issue with processing requests that have more items and
iteration should be used.