chore(build): Move Semantic Release dep to runtime
We only use `semantic-release` library for release time which is run in
a GHA build. Since this library is a bit bullish on dropping older JS
versions it often causes pending updates and forces us to move library
support up only for this reason.

By moving it to runtime we can continue supporting what we wish and just
ensure the release process is run in a modern JS version.